Change UI elements' waitXXX methods to return a value.
Change the waitXXX methods to return a value (e.g., WebElement).
This usage is not as readable as the plain wait usage, but it enables
element codes to use the located WebElement without needing to locate
it again (or reimplement the waitXXX method).

Also add a restriction on the generic type T that can be passed into
the BasePage class.